Some improvement to prepare for running the fd32 command shell
git-svn-id: https://svn.code.sf.net/p/freedos-32/code/trunk@811 85e9c19b-ca11-0410-bd92-afdf7feba87b

+++ fd32/modules/winb/advapi32.c
@@ -49,8 +49,8 @@ static LONG WINAPI fd32_imp__RegCloseKey(HKEY hKey)
 
 static char advapi32_name[] = "advapi32.dll";
 static struct symbol advapi32_symarray[] = {
-  {"RegOpenKeyExA",               (uint32_t)fd32_imp__RegOpenKeyExA},
-  {"RegCloseKey",                 (uint32_t)fd32_imp__RegCloseKey}
+  {"RegOpenKeyExA",               fd32_imp__RegOpenKeyExA},
+  {"RegCloseKey",                 fd32_imp__RegCloseKey}
 };
 static uint32_t advapi32_symnum = sizeof(advapi32_symarray)/sizeof(struct symbol);;
